chore: move packages/sendportal-core/database to  database/
