+++ .gitignore
@@ -18,4 +18,5 @@ resources/lang/vendor/
 resources/views/vendor/
 composer.lock
 phpunit.xml
+/.phpunit.cache
