+++ database/factories/TemplateFactory.php
@@ -2,22 +2,23 @@
 
 declare(strict_types=1);
 
-namespace Sendportal\Base\Database\Factories;
+namespace Database\Factories;
 
 use Illuminate\Database\Eloquent\Factories\Factory;
 use Sendportal\Base\Facades\Sendportal;
-use Sendportal\Base\Models\Tag;
+use Sendportal\Base\Models\Template;
 
-class TagFactory extends Factory
+class TemplateFactory extends Factory
 {
     /** @var string */
-    protected $model = Tag::class;
+    protected $model = Template::class;
 
     public function definition(): array
     {
         return [
+            'name' => $this->faker->word(),
             'workspace_id' => Sendportal::currentWorkspaceId(),
-            'name' => ucwords($this->faker->unique()->word())
+            'content' => '{{content}}'
         ];
     }
 }
